# src/main/resources/api.yml
openapi: 3.0.1
info:
  title: API Productos
  description: API para gestión del catálogo de productos
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Servidor local
paths:
  /api/product:
    get:
      summary: Listar catálogo de productos
      operationId: listarProductos
      tags:
        - Productos
      responses:
        '200':
          description: Lista de productos obtenida correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductoResponse'
        '204':
          description: "Sin resultados"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Registrar nuevo producto
      operationId: registrarProducto
      tags:
        - Productos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductoRequest'
      responses:
        '200':
          description: Producto registrado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoResponse'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/products/{id}:
    get:
      summary: Obtener producto por ID
      operationId: obtenerProductoPorId
      tags:
        - Productos
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador único del producto
          schema:
            type: string
      responses:
        '200':
          description: Producto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoResponse'
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:

    ProductoRequest:
      type: object
      required:
        - nombre
        - precio
        - categoria
      properties:
        nombre:
          type: string
          example: Laptop Lenovo ThinkPad
        descripcion:
          type: string
          example: Laptop empresarial con 16GB RAM
        precio:
          type: number
          format: double
          example: 1200.50
        categoria:
          type: string
          example: Tecnología
        stock:
          type: integer
          format: int32
          example: 10

    ProductoResponse:
      type: object
      properties:
        id:
          type: string
          example: "12345"
        nombre:
          type: string
          example: Laptop Lenovo ThinkPad
        descripcion:
          type: string
          example: Laptop empresarial con 16GB RAM
        precio:
          type: number
          format: double
          example: 1200.50
        categoria:
          type: string
          example: Tecnología
        stock:
          type: integer
          format: int32
          example: 10
        fechaCreacion:
          type: string
          format: date-time
          example: 2026-02-18T10:30:00Z

    ErrorResponse:
      type: object
      properties:
        codigo:
          type: string
          example: ERROR_NEGOCIO
        mensaje:
          type: string
          example: El producto no fue encontrado
        detalle:
          type: string
          example: No existe un producto con el id proporcionado